#328 - addressed issues in the pull request, including refactoring tests, renaming variables, and moving loading logic to the parent component
