+++ src/components/change-requests/new-change-request/new-change-request-page/common-form-fields/common-form-fields.tsx
@@ -7,10 +7,10 @@ import { Form } from 'react-bootstrap';
 import { wbsPipe } from '../../../../../shared/pipes';
 import { Project, WorkPackage } from 'utils';
 import './common-form-fields.module.css';
-import { CR_Type } from '@prisma/client';
+import { ChangeRequestType } from 'utils/src';
 
 interface CommonFormFieldsProp {
-  setType: React.Dispatch<React.SetStateAction<CR_Type>>,
+  setType: React.Dispatch<React.SetStateAction<ChangeRequestType>>,
   projects: Project[],
   workPkgs: WorkPackage[],
   handleChange: (e: any) => void,
@@ -46,9 +46,9 @@ const CommonFormFields: React.FC<CommonFormFieldsProp> = ({projects, workPkgs, s
         </Form.Control>
       </div>
       <div className={'px-4'}>
-        Form Type
+        Type
         <Form.Control as="select" custom onChange={handleType} data-testid="type" name="type">
-          {Object.values(CR_Type).map((t) => (
+          {Object.values(ChangeRequestType).map((t) => (
             <option value={t}>{t}</option>
           ))}
         </Form.Control>
