+++ src/components/change-requests/new-change-request/new-change-request-page/stage-gate-form-fileds/stage-gate-form-fields.test.tsx
@@ -6,29 +6,32 @@
 import { render, screen } from '../../../../../test-support/test-utils';
 import StageGateFormFields from './stage-gate-form-fields';
 
+const requiredInquiry = 'Who is Required for Design Review?';
+const leftoverBudgetStr = 'Leftover Budget';
+const doneInquiry = 'Is everything done?';
 /**
  * Sets up the component under test with the desired values and renders it.
  */
-const renderComponent = () => {
-  return render(<StageGateFormFields />);
+const renderComponent = (handleChange: jest.Mock<any, any>) => {
+  return render(<StageGateFormFields handleChange={handleChange}/>);
 };
 
 describe('new stage gate form fields', () => {
   it('renders the design reviewer form field', () => {
-    renderComponent();
+    renderComponent(jest.fn());
 
-    expect(screen.getByText('Who is Required for Design Review?')).toBeInTheDocument();
+    expect(screen.getByText(requiredInquiry)).toBeInTheDocument();
   });
 
   it('renders the Leftover Budget form field', () => {
-    renderComponent();
+    renderComponent(jest.fn());
 
-    expect(screen.getByText('Leftover Budget')).toBeInTheDocument();
+    expect(screen.getByText(leftoverBudgetStr)).toBeInTheDocument();
   });
 
   it('renders the status form field', () => {
-    renderComponent();
+    renderComponent(jest.fn());
 
-    expect(screen.getByText('Is everything done?')).toBeInTheDocument();
+    expect(screen.getByText(doneInquiry)).toBeInTheDocument();
   });
 });
